feat(demo): reorganize TAM edge function and enhance thank-you hero styling
- Move tamDomains.ts to lib/ subdirectory to prevent Netlify from treating it as an edge function
- Update import path in isTargetAccount.ts to reference relocated module
- Clarify documentation that tamDomains is a shared supporting module, not an edge function
- Widen video column in thank-you hero from 600px to 648px for better visual balance
- Increase heading size from clamp(29px, 3.4vw, 50px) to clamp(34px, 5vw, 58px) for stronger visual hierarchy
- Add flexbox centering to non-TAM thank-you hero to vertically center content within viewport
- Reset min-height to auto on responsive breakpoint to allow stacked content to flow naturally
- Update grid layout comments to reflect new column width and styling rationale
