+++ src/pages/demo/thankyou.module.css
@@ -36,12 +36,25 @@
     radial-gradient(125% 110% at 0% 0%, rgba(0, 46, 44, 0.6) 0%, rgba(0, 46, 44, 0) 58%),
     radial-gradient(110% 95% at 100% 100%, rgba(165, 240, 220, 0.26) 0%, rgba(165, 240, 220, 0) 55%);
 }
+
+/* Non-TAM (video) page fills the viewport below the navbar + banner bar and
+   vertically centers its content (the `.xl-container` wrapper). */
+.tyHero:not(.tyHeroTam) {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  min-height: calc(
+    100vh - var(--navbar-height, 84px) - var(--announcment-bar-height, 50px)
+  );
+}
+
 /* Width + inline padding come from the shared `.xl-container` wrapper (as on the
    demo page); this only owns the two-column grid and vertical rhythm. */
 .tyHeroInner {
   position: relative;
   display: grid;
-  grid-template-columns: minmax(0, 1fr) 600px;
+  /* Video column widened ~8% (600px → 648px). */
+  grid-template-columns: minmax(0, 1fr) 648px;
   column-gap: 72px;
   align-items: center;
   padding: 56px 0 64px;
@@ -69,11 +82,11 @@
   flex: none;
   animation: bdCheckPop 0.55s 0.1s var(--ease-emphasis) both;
 }
-/* Heading + lead scale mirrors the demo hero (`--font-display`, ~29→50px). */
+/* Heading tops out at 58px on desktop, scaling down on narrower screens. */
 .tyH1 {
   font-family: var(--font-display);
   font-weight: 800;
-  font-size: clamp(29px, 3.4vw, 50px);
+  font-size: clamp(34px, 5vw, 58px);
   line-height: 1.08;
   letter-spacing: -0.02em;
   margin: 0;
@@ -220,6 +233,10 @@
 
 /* ---- responsive --------------------------------------------------------- */
 @media (max-width: 980px) {
+  /* Stacked content can exceed the viewport, so let the section grow naturally. */
+  .tyHero:not(.tyHeroTam) {
+    min-height: auto;
+  }
   .tyHeroInner {
     grid-template-columns: 1fr;
     row-gap: 40px;
